From Copilot, switch the usage of `byte` with `uint32`:
> Using a Go `byte`/`uint8` for a JSON field means invalid values
> (negative, non-integer, >255) will fail during JSON unmarshal,
> before `handleEmit` can wrap the error as `ErrClientError`. To
> keep error handling consistent and provide clearer client errors,
> consider unmarshalling `emitOnly` as an `int` (or `uint32`) and
> doing explicit validation in `handleEmit`.

+++ internal/api/proto.go
@@ -967,7 +967,7 @@ type EmitParams struct {
 	Snapshot         SnapshotID          `json:"snapshot"`
 	Project          ProjectID           `json:"project"`
 	TargetSourceFile *DocumentIdentifier `json:"targetSourceFile,omitempty"`
-	EmitOnly         byte                `json:"emitOnly"`
+	EmitOnly         uint32              `json:"emitOnly"`
 }
 
 // EmitResponse is the response for the emit method.
